ward-security-classifier/requirements.txt
torch>=2.4.0
transformers>=4.45.0
peft>=0.13.0
trl>=0.11.0
datasets>=3.0.0
accelerate>=0.34.0
bitsandbytes>=0.44.0
safetensors>=0.4.5
scipy
pyyaml
